+++ jkube-kit/enricher/api/src/main/java/org/eclipse/jkube/kit/enricher/api/util/Constants.java
@@ -18,7 +18,15 @@ package org.eclipse.jkube.kit.enricher.api.util;
 // TODO-F8SPEC : Should be move the to AppCatalog mojo and must not be in the general available util package
 // Also consider whether the Constants class pattern makes (should probably change to real enums ???)
 public class Constants {
+    /**
+     * @deprecated Use annotations with <code>jkube.eclipse.org</code> prefix instead
+     */
+    @Deprecated
     public static final String RESOURCE_SOURCE_URL_ANNOTATION = "maven.jkube.io/source-url";
+    /**
+     * @deprecated Use annotations with <code>jkube.eclipse.org</code> prefix instead
+     */
+    @Deprecated
     public static final String RESOURCE_APP_CATALOG_ANNOTATION = "maven.jkube.io/app-catalog";
 
     private Constants() { }